Needs an usize not an u32.

+++ src/log.rs
@@ -1,4 +1,4 @@
 #[no_mangle]
-pub extern "C" fn bochscpu_log_set_level(level: u32) {
+pub extern "C" fn bochscpu_log_set_level(level: usize) {
     stderrlog::new().verbosity(level).init().unwrap();
 }
